Documentación y corrección de errores.
Se corrigieron errores con respecto a mensajes, fondos de ventanas, conexión sql, optimización, procedimientos sql.

Se incluye el término de la documentación en formato PDF.

+++ Archivos SQL/PUNTO18ARQUITECTOS.sql
@@ -61,7 +61,7 @@ create table Usuarios
 	idUsuario int identity(1,1),
 	Nombre varchar(40)not null,	
 	direccion varchar(20),
-	telefono int,
+	telefono bigint,
 	idCiudad int ,
 	primary key(idUsuario),
 	foreign key (idCiudad) references Ciudades(idCiudad),
